% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-features.R
\name{apply_collapse_rows}
\alias{apply_collapse_rows}
\title{Apply collapse rows to kable object}
\usage{
apply_collapse_rows(kable_obj, collapse_spec)
}
\arguments{
\item{kable_obj}{A kable object.}

\item{collapse_spec}{A t2f_collapse object.}
}
\value{
Modified kable object with collapsed rows.
}
\description{
Apply collapse rows to kable object
}
\keyword{internal}
