% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-features.R
\name{apply_footnotes}
\alias{apply_footnotes}
\title{Apply footnotes to kable object}
\usage{
apply_footnotes(kable_obj, footnote_spec)
}
\arguments{
\item{kable_obj}{A kable object.}

\item{footnote_spec}{A t2f_footnote object.}
}
\value{
Modified kable object with footnotes.
}
\description{
Apply footnotes to kable object
}
\keyword{internal}
