% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-features.R
\name{apply_header_above}
\alias{apply_header_above}
\title{Apply spanning headers to kable object}
\usage{
apply_header_above(kable_obj, header_spec)
}
\arguments{
\item{kable_obj}{A kable object.}

\item{header_spec}{A t2f_header object or list of them.}
}
\value{
Modified kable object with spanning headers.
}
\description{
Apply spanning headers to kable object
}
\keyword{internal}
