% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{apply_single_format}
\alias{apply_single_format}
\title{Apply a single formatting specification}
\usage{
apply_single_format(kable_obj, df, fmt)
}
\arguments{
\item{kable_obj}{A kableExtra table object.}

\item{df}{The original data frame.}

\item{fmt}{A t2f_format object.}
}
\value{
The modified kableExtra table object.
}
\description{
Apply a single formatting specification
}
\keyword{internal}
