% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_single_logical}
\alias{assert_single_logical}
\title{Assert parameter is a single logical value}
\usage{
assert_single_logical(x, name)
}
\arguments{
\item{x}{Value to check.}

\item{name}{Parameter name for error message.}
}
\description{
Assert parameter is a single logical value
}
\keyword{internal}
