% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_single_string}
\alias{assert_single_string}
\title{Assert parameter is a single character string}
\usage{
assert_single_string(x, name)
}
\arguments{
\item{x}{Value to check.}

\item{name}{Parameter name for error message.}
}
\description{
Assert parameter is a single character string
}
\keyword{internal}
