% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_string_or_null}
\alias{assert_string_or_null}
\title{Assert parameter is a single character string or NULL}
\usage{
assert_string_or_null(x, name)
}
\arguments{
\item{x}{Value to check.}

\item{name}{Parameter name for error message.}
}
\description{
Assert parameter is a single character string or NULL
}
\keyword{internal}
