% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_type}
\alias{assert_type}
\title{Assert parameter meets type requirements}
\usage{
assert_type(x, name, type, allow_null = FALSE, single = TRUE)
}
\arguments{
\item{x}{Value to check.}

\item{name}{Parameter name for error message.}

\item{type}{Expected type ("logical", "character").}

\item{allow_null}{Logical. Allow NULL values.}

\item{single}{Logical. Require single value.}
}
\description{
Assert parameter meets type requirements
}
\keyword{internal}
