% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab2fig.R
\name{auto_align}
\alias{auto_align}
\title{Auto-detect column alignment based on data types}
\usage{
auto_align(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A character vector of alignments ("l", "c", or "r").
}
\description{
Auto-detect column alignment based on data types
}
\keyword{internal}
