% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab2fig.R
\name{babel}
\alias{babel}
\title{Create a babel package specification for language support}
\usage{
babel(language)
}
\arguments{
\item{language}{Language code (e.g., "spanish", "french", "german")}
}
\value{
Character string with LaTeX babel package specification
}
\description{
Create a babel package specification for language support
}
