% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-methods.R
\name{broom-methods}
\alias{broom-methods}
\title{Broom-Based S3 Methods for Extended Object Support}
\description{
S3 methods that use the broom package to convert statistical
objects to tidy data frames before generating LaTeX tables. These methods
extend zztab2fig's support to survival models, time series, clustering,
and other object types.
}
