% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline.R
\name{build_inline_latex}
\alias{build_inline_latex}
\title{Build LaTeX code for inline table}
\usage{
build_inline_latex(
  path,
  width,
  height,
  align,
  caption,
  caption_short,
  label,
  caption_position
)
}
\arguments{
\item{path}{Path to the PDF file.}

\item{width}{Width specification.}

\item{height}{Height specification.}

\item{align}{Alignment (left, center, right).}

\item{caption}{Caption text.}

\item{caption_short}{Short caption for LoT.}

\item{label}{LaTeX label.}

\item{caption_position}{Above or below.}
}
\value{
Character string with LaTeX code.
}
\description{
Build LaTeX code for inline table
}
\keyword{internal}
