% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms-methods.R
\name{build_rms_model_stats}
\alias{build_rms_model_stats}
\title{Build model statistics rows for rms models}
\usage{
build_rms_model_stats(models, digits)
}
\description{
Build model statistics rows for rms models
}
\keyword{internal}
