% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_system_command}
\alias{check_system_command}
\title{Check if a system command is available}
\usage{
check_system_command(cmd)
}
\arguments{
\item{cmd}{Command name to check.}
}
\value{
Logical indicating if command is available.
}
\description{
Check if a system command is available
}
\keyword{internal}
