% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-formats.R
\name{command_exists}
\alias{command_exists}
\title{Check if a system command is available}
\usage{
command_exists(cmd)
}
\arguments{
\item{cmd}{Command name to check.}
}
\value{
Logical indicating if command is available.
}
\description{
Check if a system command is available
}
\keyword{internal}
