% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzdataframe2graphic.R
\name{compile_latex}
\alias{compile_latex}
\title{Compile a LaTeX file to PDF}
\usage{
compile_latex(tex_file, sub_dir)
}
\arguments{
\item{tex_file}{Path to the LaTeX file.}

\item{sub_dir}{Directory where the PDF will be generated.}
}
\description{
Compile a LaTeX file to PDF
}
