% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{compute_cache_hash}
\alias{compute_cache_hash}
\title{Compute hash for cache key}
\usage{
compute_cache_hash(
  df,
  scolor = NULL,
  extra_packages = NULL,
  document_class = NULL,
  caption = NULL,
  label = NULL,
  align = NULL,
  longtable = FALSE,
  theme = NULL
)
}
\arguments{
\item{df}{The data frame being processed.}

\item{scolor}{Row shading color.}

\item{extra_packages}{Extra LaTeX packages.}

\item{document_class}{Document class.}

\item{caption}{Table caption.}

\item{label}{Table label.}

\item{align}{Column alignment.}

\item{longtable}{Longtable setting.}

\item{theme}{Theme object or name.}
}
\value{
A character string hash.
}
\description{
Compute a hash of the inputs to t2f for use as a cache key.
}
\keyword{internal}
