% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab2fig.R
\name{create_latex_template}
\alias{create_latex_template}
\title{Create LaTeX document template with specified packages}
\usage{
create_latex_template(document_class = "article", extra_packages = NULL)
}
\arguments{
\item{document_class}{LaTeX document class}

\item{extra_packages}{List of package specifications or character strings}
}
\value{
Character string with complete LaTeX document preamble
}
\description{
Create LaTeX document template with specified packages
}
