% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms-methods.R
\name{extract_rms_anova}
\alias{extract_rms_anova}
\title{Extract ANOVA-style output from rms objects}
\usage{
extract_rms_anova(x, digits = 3)
}
\arguments{
\item{x}{An rms model object.}

\item{digits}{Number of decimal places.}
}
\value{
A data frame with ANOVA results.
}
\description{
Internal helper to extract ANOVA (chunk test) results from
rms model objects.
}
\keyword{internal}
