% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab2fig.R
\name{fontspec}
\alias{fontspec}
\title{Create fontspec package specification for custom fonts}
\usage{
fontspec(main_font = NULL, sans_font = NULL, mono_font = NULL)
}
\arguments{
\item{main_font}{Main font family name}

\item{sans_font}{Sans-serif font family name}

\item{mono_font}{Monospace font family name}
}
\value{
Character vector with LaTeX fontspec package specifications
}
\description{
Create fontspec package specification for custom fonts
}
