% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_pvalue}
\alias{format_pvalue}
\title{Format p-values for display}
\usage{
format_pvalue(p, digits = 3)
}
\arguments{
\item{p}{Numeric vector of p-values.}

\item{digits}{Number of decimal places.}
}
\value{
Character vector of formatted p-values.
}
\description{
Format p-values for display
}
\keyword{internal}
