% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{format_with_stars}
\alias{format_with_stars}
\title{Format estimates with significance stars}
\usage{
format_with_stars(estimates, pvals, stars, digits)
}
\description{
Format estimates with significance stars
}
\keyword{internal}
