% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab2fig.R
\name{geometry}
\alias{geometry}
\title{Create a geometry package specification}
\usage{
geometry(margin = NULL, paper = NULL, landscape = FALSE, ...)
}
\arguments{
\item{margin}{Page margin specification (e.g., "5mm")}

\item{paper}{Paper size specification (e.g., "a4paper")}

\item{landscape}{Logical. If TRUE, sets landscape orientation.}

\item{...}{Additional geometry options}
}
\value{
Character string with LaTeX geometry package specification
}
\description{
Create a geometry package specification
}
