% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{get_builtin_theme}
\alias{get_builtin_theme}
\title{Get a built-in theme by name}
\usage{
get_builtin_theme(name)
}
\arguments{
\item{name}{Character string. Name of the built-in theme.}
}
\value{
A t2f_theme object.
}
\description{
Get a built-in theme by name
}
\keyword{internal}
