% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-engine.R, R/themes.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator (for knitr engine)}
\usage{
x \%||\% y

x \%||\% y
}
\description{
Null-coalescing operator (for knitr engine)

Null-coalescing operator
}
\keyword{internal}
