% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex-include.R
\name{latex-include}
\alias{latex-include}
\title{LaTeX Figure Inclusion Helpers}
\description{
Functions to generate LaTeX code for including t2f-generated
tables as figures in LaTeX documents. Useful in R Markdown with
\code{results='asis'}.
}
