% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{print.t2f_format}
\alias{print.t2f_format}
\title{Print method for t2f_format objects}
\usage{
\method{print}{t2f_format}(x, ...)
}
\arguments{
\item{x}{A t2f_format object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns x.
}
\description{
Print method for t2f_format objects
}
