% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab2fig.R
\name{protect_siunitx_headers}
\alias{protect_siunitx_headers}
\title{Protect siunitx column headers with braces}
\usage{
protect_siunitx_headers(latex_table, siunitx_cols)
}
\arguments{
\item{latex_table}{The LaTeX table string (or kable object).}

\item{siunitx_cols}{Integer vector of column indices with siunitx alignment.}
}
\value{
Modified LaTeX table with protected headers.
}
\description{
For siunitx S columns, non-numeric content in the header must
be wrapped in braces to prevent siunitx from trying to parse it as a
number. This function post-processes the LaTeX table to add braces around
headers in S columns.
}
\keyword{internal}
