% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-engine.R
\name{register_t2f_engine}
\alias{register_t2f_engine}
\title{Register the t2f knitr engine}
\usage{
register_t2f_engine()
}
\value{
Invisible NULL.
}
\description{
Register the t2f engine with knitr. This is called
automatically when the package is loaded.
}
