% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{require_package}
\alias{require_package}
\title{Check that a package is available}
\usage{
require_package(pkg, for_what = "this object type")
}
\arguments{
\item{pkg}{Package name.}

\item{for_what}{Description of what requires this package.}
}
\description{
Check that a package is available
}
\keyword{internal}
