% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex-include.R
\name{resolve_pdf_path}
\alias{resolve_pdf_path}
\title{Resolve PDF path for inclusion}
\usage{
resolve_pdf_path(path)
}
\arguments{
\item{path}{Input path (may or may not have _cropped.pdf)}
}
\value{
Resolved path with _cropped.pdf
}
\description{
Resolve PDF path for inclusion
}
\keyword{internal}
