% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{retrieve_from_cache}
\alias{retrieve_from_cache}
\title{Retrieve from cache}
\usage{
retrieve_from_cache(cache_path, target_path)
}
\arguments{
\item{cache_path}{Path to the cached file.}

\item{target_path}{Target path for the output.}
}
\value{
Logical indicating success.
}
\description{
Copy a cached file to the target location.
}
\keyword{internal}
