% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_numeric_cols}
\alias{round_numeric_cols}
\title{Round numeric columns in a data frame}
\usage{
round_numeric_cols(df, digits)
}
\arguments{
\item{df}{A data frame.}

\item{digits}{Number of decimal places.}
}
\value{
Data frame with rounded numeric columns.
}
\description{
Round numeric columns in a data frame
}
\keyword{internal}
