% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zztab2fig.R
\name{sanitize_column_names}
\alias{sanitize_column_names}
\title{Sanitize column names to be valid in both R and LaTeX}
\usage{
sanitize_column_names(colnames)
}
\arguments{
\item{colnames}{A character vector of column names.}
}
\value{
A sanitized character vector of column names.
}
\description{
Sanitize column names to be valid in both R and LaTeX
}
