% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzdataframe2graphic.R
\name{sanitize_filename}
\alias{sanitize_filename}
\title{Sanitize filenames to be file-system safe}
\usage{
sanitize_filename(filename)
}
\arguments{
\item{filename}{A character string.}
}
\value{
A sanitized character string.
}
\description{
Sanitize filenames to be file-system safe
}
