% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab2fig.R
\name{sanitize_table_cells}
\alias{sanitize_table_cells}
\title{Sanitize table cells to be LaTeX-safe}
\usage{
sanitize_table_cells(cells)
}
\arguments{
\item{cells}{A character vector of table cell values.}
}
\value{
A sanitized character vector of table cell values.
}
\description{
Sanitize table cells to be LaTeX-safe
}
