% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{store_in_cache}
\alias{store_in_cache}
\title{Store output in cache}
\usage{
store_in_cache(source_path, hash)
}
\arguments{
\item{source_path}{Path to the generated file.}

\item{hash}{Cache key hash.}
}
\value{
Path to the cached file.
}
\description{
Copy the generated output to the cache directory.
}
\keyword{internal}
