% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{t2f.anova}
\alias{t2f.anova}
\alias{t2f.aov}
\title{t2f method for ANOVA objects}
\usage{
\method{t2f}{anova}(x, digits = 3, ...)

\method{t2f}{aov}(x, digits = 3, ...)
}
\arguments{
\item{x}{An anova or aov object.}

\item{digits}{Number of decimal places.}

\item{...}{Additional arguments passed to t2f.default.}
}
\value{
Invisibly returns the path to the generated PDF.
}
\description{
t2f method for ANOVA objects

t2f method for aov objects
}
