% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{t2f.lm}
\alias{t2f.lm}
\title{t2f method for linear models}
\usage{
\method{t2f}{lm}(
  x,
  digits = 3,
  include = c("estimate", "std.error", "statistic", "p.value"),
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{x}{An lm object.}

\item{digits}{Number of decimal places for coefficients.}

\item{include}{Character vector of statistics to include. Options:
"estimate", "std.error", "statistic", "p.value", "conf.int".}

\item{conf.level}{Confidence level for confidence intervals.}

\item{...}{Additional arguments passed to t2f.default.}
}
\value{
Invisibly returns the path to the generated PDF.
}
\description{
t2f method for linear models
}
\examples{
\dontrun{
model <- lm(mpg ~ cyl + hp + wt, data = mtcars)
t2f(model, filename = "regression_table")
}

}
