% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{t2f}
\alias{t2f}
\title{Convert objects to LaTeX tables}
\usage{
t2f(x, ...)
}
\arguments{
\item{x}{An object to convert to a LaTeX table.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Invisibly returns the path to the generated PDF file.
}
\description{
Generic function for converting R objects to LaTeX tables
and generating PDF output.
}
\examples{
\dontrun{
# Data frame
t2f(mtcars)

# Linear model
model <- lm(mpg ~ cyl + hp, data = mtcars)
t2f(model)

# t-test
t2f(t.test(mtcars$mpg, mu = 20))
}

}
