% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{t2f_cache_info}
\alias{t2f_cache_info}
\title{Get cache statistics}
\usage{
t2f_cache_info()
}
\value{
A list with cache statistics.
}
\description{
Report on the current state of the t2f cache.
}
\examples{
\dontrun{
t2f_cache_info()
}

}
