% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{t2f_color_row}
\alias{t2f_color_row}
\title{Color specific rows}
\usage{
t2f_color_row(rows, background)
}
\arguments{
\item{rows}{Integer vector of row indices (1-based, excluding header).}

\item{background}{LaTeX color for row background.}
}
\value{
A t2f_format object with specified background for rows.
}
\description{
Convenience function to apply background color to rows.
}
\examples{
\dontrun{
t2f_color_row(1, "gray!20")
t2f_color_row(c(1, 3, 5), "blue!10")
}

}
