% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{t2f_list_themes}
\alias{t2f_list_themes}
\title{List available themes}
\usage{
t2f_list_themes(builtin_only = FALSE)
}
\arguments{
\item{builtin_only}{Logical. If TRUE, only list built-in themes. Default
FALSE.}
}
\value{
Character vector of available theme names.
}
\description{
List all available themes, including both built-in themes and
registered custom themes.
}
\examples{
t2f_list_themes()
t2f_list_themes(builtin_only = TRUE)

}
