% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{t2f_theme_apa}
\alias{t2f_theme_apa}
\title{APA theme}
\usage{
t2f_theme_apa()
}
\value{
A t2f_theme object.
}
\description{
Theme following American Psychological Association (APA 7th
edition) table guidelines. Characterized by no row shading, bold headers,
and booktabs-style horizontal rules.
}
