% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{t2f_theme_get}
\alias{t2f_theme_get}
\title{Get the current global t2f theme}
\usage{
t2f_theme_get()
}
\value{
The current t2f_theme object, or NULL if no theme is set.
}
\description{
Retrieve the currently active global theme.
}
