% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{t2f_theme_minimal}
\alias{t2f_theme_minimal}
\title{Minimal theme}
\usage{
t2f_theme_minimal()
}
\value{
A t2f_theme object.
}
\description{
A clean, minimal theme with light gray shading and no extra
formatting. Suitable for general-purpose use.
}
