% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{t2f_theme_unregister}
\alias{t2f_theme_unregister}
\title{Unregister a custom theme}
\usage{
t2f_theme_unregister(name)
}
\arguments{
\item{name}{Character string. Name of the theme to unregister.}
}
\value{
Invisibly returns TRUE if the theme was removed, FALSE if it was not
found.
}
\description{
Remove a previously registered custom theme.
}
\examples{
\dontrun{
t2f_theme_unregister("jama")
}

}
