% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zztab2fig.R
\name{crop_pdf}
\alias{crop_pdf}
\title{Crop a PDF file}
\usage{
crop_pdf(input_pdf, output_pdf)
}
\arguments{
\item{input_pdf}{Path to the input PDF file.}

\item{output_pdf}{Path to the output cropped PDF file.}
}
\description{
Crop a PDF file
}
