% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{get_builtin_theme}
\alias{get_builtin_theme}
\title{Get a theme by name}
\usage{
get_builtin_theme(name)
}
\arguments{
\item{name}{Character string. Name of the theme.}
}
\value{
A t2f_theme object.
}
\description{
Look up a theme by name, checking both built-in themes and
registered custom themes.
}
\keyword{internal}
