% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-methods.R
\name{round_numeric_cols}
\alias{round_numeric_cols}
\title{Round numeric columns in a data frame}
\usage{
round_numeric_cols(df, digits)
}
\description{
Round numeric columns in a data frame
}
\keyword{internal}
