% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline.R
\name{t2f_coef}
\alias{t2f_coef}
\title{Quick coefficient table from model}
\usage{
t2f_coef(
  model,
  width = "3in",
  align = "left",
  digits = 3,
  stars = TRUE,
  theme = "minimal",
  caption = NULL,
  caption_short = NULL,
  label = NULL,
  ...
)
}
\arguments{
\item{model}{A fitted model object (lm, glm, coxph, etc.).}

\item{width}{Figure width. Default "3in".}

\item{align}{Alignment. Default "left".}

\item{digits}{Number of decimal places.}

\item{stars}{Show significance stars. Default TRUE.}

\item{theme}{Table theme. Default "minimal".}

\item{caption}{Optional table caption.}

\item{caption_short}{Short caption for List of Tables. Defaults to NULL.}

\item{label}{LaTeX label for cross-referencing. Defaults to NULL.}

\item{...}{Additional arguments passed to t2f_inline().}
}
\value{
Same as t2f_inline().
}
\description{
Convenience function to generate an inline coefficient
table from a regression model with sensible defaults.
}
\examples{
\dontrun{
model <- lm(mpg ~ cyl + hp + wt, data = mtcars)
t2f_coef(model)
}

}
