% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{t2f_list_themes}
\alias{t2f_list_themes}
\title{List available built-in themes}
\usage{
t2f_list_themes()
}
\value{
Character vector of available theme names.
}
\description{
List available built-in themes
}
